part4 done

+++ osa4/blogilista/controllers/blogs.js
@@ -1,18 +1,18 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
-  .find({}).populate('user', { username: 1, name: 1 })
+    .find({}).populate('user', { username: 1, name: 1 })
 
   response.json(blogs)
-  })
-  
-blogsRouter.post('/', async (request, response) => {
+})
+
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
 
-  const user = await User.findById(body.userId)
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -21,7 +21,7 @@ blogsRouter.post('/', async (request, response) => {
     likes: body.likes || 0,
     user: user._id
   })
-  
+
   const savedBlog = await blog.save()
 
   user.blogs = user.blogs.concat(savedBlog._id)
@@ -32,16 +32,24 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
-    if (blog) {
-      response.json(blog)
-    } else {
-      response.status(404).json(savedBlog)
-    }
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).json({ error: 'blog removed from server' })
+  }
 })
-  
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  const user = request.user
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'invalid user' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
